Feature: Added support of scheduling devices through RainMaker app.
Schedules allow you to automate a device by setting it to trigger an action at the same time on a specified day or days of the week.
App provides a separate flow for accessing Schedule feature in Home Screen under the Schedules tab.
List of schedule that are already added for a device is fetched from the cloud and displayed on the app.
User is given option to add a new schedule, enable/disable, remove or edit an existing schedule.
Only devices that support scheduling will be available on the action list for automation.
Scheduling feature is optional and can be disabled from build settings(steps are given in README).

Update: App is now compatible to run on Xcode 12.0 and iOS 14.