Feature: Added support of assisted claiming through RainMaker App.
Claiming is the process by which the ESP32-S2/ESP32 gets certificate from the ESP RainMaker Claiming Service.
App determines from the version information of the device whether claiming is supported/required.
If it is required, app initiates series of information exchange of data and certificates between device, app and claiming service to perform claiming.
A separate screen is added to show progress of claiming, error and resolution.

Note: This feature is supported only for ESP32 BLE Provisioning.