Feature: Added support for interacting with device(s) over local network
Added support to interact with RainMaker devices using the "ESP Local Control" feature of ESP-IDF.
The app searches for the appropriate services on the local Wi-Fi network using Bonjour(mDNS).
It then resolves the found services and filters the devices associated with the user.
Then communication with the filtered device(s) occurs over the local network using HTTP.